bind question deletion to user_id

+++ server.py
@@ -142,9 +142,12 @@ def route_new_answer(question_id):
 
 @app.route('/question/<question_id>/delete')
 def route_delete_question(question_id):
-    data_manager.delete_question(question_id)
-
-    return redirect(url_for('route_list'))
+    if 'user_id' in session:
+        user_id_for_question = data_manager.get_user_id_for_question(question_id)
+        if session['user_id'] == user_id_for_question:
+            data_manager.delete_question(question_id)
+            return redirect(url_for('route_index'))
+    return redirect(url_for('display_question_and_answers', question_id=question_id))
 
 
 @app.route('/question/<question_id>/<answer_id>/delete')
